Organized dashboard pages to use local components folder, changed named exports to default exports in modules that only export one function
